# docs/evidencias/openapi.json
{"openapi":"3.1.0","info":{"title":"API - Educacao Continuada Gamificada","description":"Plataforma de educacao continuada com gamificacao (XP, niveis e badges). Projeto academico desenvolvido com TDD/BDD (ATDD) em Spring Boot.","contact":{"name":"Mathues de Luzia Souza, Emanuel Ronaldo Gomes de Souza, Luis Henrique Telo Ladeira Mota"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Cursos","description":"Catalogo de cursos/trilhas de educacao continuada"},{"name":"Matriculas","description":"Matricula de alunos em cursos e conclusao de trilhas (concede XP)"},{"name":"Usuarios","description":"Cadastro de alunos e ranking de gamificacao"}],"paths":{"/api/usuarios":{"get":{"tags":["Usuarios"],"summary":"Lista todos os alunos cadastrados","operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UsuarioDTO"}}}}}}},"post":{"tags":["Usuarios"],"summary":"Cadastra um novo aluno","operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioDTO"}}}}}}},"/api/matriculas":{"get":{"tags":["Matriculas"],"summary":"Lista as matriculas de um aluno","operationId":"listarPorUsuario","parameters":[{"name":"usuarioId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MatriculaDTO"}}}}}}},"post":{"tags":["Matriculas"],"summary":"Matricula um aluno em um curso","operationId":"matricular","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatriculaCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatriculaDTO"}}}}}}},"/api/matriculas/{id}/concluir":{"post":{"tags":["Matriculas"],"summary":"Conclui uma matricula e concede XP ao aluno","operationId":"concluir","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatriculaDTO"}}}}}}},"/api/cursos":{"get":{"tags":["Cursos"],"summary":"Lista todos os cursos disponiveis","operationId":"listar_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CursoDTO"}}}}}}},"post":{"tags":["Cursos"],"summary":"Cadastra um novo curso","operationId":"criar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CursoCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursoDTO"}}}}}}},"/api/usuarios/{id}":{"get":{"tags":["Usuarios"],"summary":"Busca um aluno pelo id","operationId":"buscarPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioDTO"}}}}}}},"/api/usuarios/ranking":{"get":{"tags":["Usuarios"],"summary":"Lista o ranking de alunos ordenado por XP (maior para o menor)","operationId":"ranking","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RankingItemDTO"}}}}}}}},"/api/cursos/{id}":{"get":{"tags":["Cursos"],"summary":"Busca um curso pelo id","operationId":"buscarPorId_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursoDTO"}}}}}}}},"components":{"schemas":{"UsuarioCreateDTO":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"email":{"type":"string","minLength":1}}},"UsuarioDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"email":{"type":"string"},"xp":{"type":"integer","format":"int32"},"nivel":{"type":"integer","format":"int32"},"criadoEm":{"type":"string","format":"date-time"}}},"MatriculaCreateDTO":{"type":"object","properties":{"usuarioId":{"type":"integer","format":"int64"},"cursoId":{"type":"integer","format":"int64"}},"required":["cursoId","usuarioId"]},"MatriculaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"usuarioId":{"type":"integer","format":"int64"},"nomeUsuario":{"type":"string"},"cursoId":{"type":"integer","format":"int64"},"tituloCurso":{"type":"string"},"dataMatricula":{"type":"string","format":"date-time"},"concluida":{"type":"boolean"},"dataConclusao":{"type":"string","format":"date-time"}}},"CursoCreateDTO":{"type":"object","properties":{"titulo":{"type":"string","minLength":1},"descricao":{"type":"string"},"cargaHoraria":{"type":"integer","format":"int32"},"xpRecompensa":{"type":"integer","format":"int32"}}},"CursoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"descricao":{"type":"string"},"cargaHoraria":{"type":"integer","format":"int32"},"xpRecompensa":{"type":"integer","format":"int32"}}},"RankingItemDTO":{"type":"object","properties":{"posicao":{"type":"integer","format":"int32"},"usuarioId":{"type":"integer","format":"int64"},"nome":{"type":"string"},"xp":{"type":"integer","format":"int32"},"nivel":{"type":"integer","format":"int32"}}}}}}